@@ -48,65 +57,67 @@
   },
   "funding": "https://github.com/moleculerjs/moleculer?sponsor=1",
   "bin": {
-    "moleculer-runner": "./bin/moleculer-runner.js"
+    "moleculer-runner": "./bin/moleculer-runner.js",
+    "moleculer-runner-esm": "./bin/moleculer-runner.mjs"
   },
   "author": "MoleculerJS",
   "license": "MIT",
   "devDependencies": {
     "@icebob/node-memwatch": "^2.1.0",
-    "@sinonjs/fake-timers": "^8.1.0",
+    "@sinonjs/fake-timers": "^9.1.2",
     "@types/bunyan": "^1.8.8",
-    "@types/ioredis": "^4.28.7",
-    "@types/node": "^16.11.1",
-    "@types/pino": "^6.3.12",
+    "@types/ioredis": "^4.28.10",
+    "@types/node": "^17.0.24",
+    "@types/pino": "^7.0.5",
     "amqplib": "^0.8.0",
-    "avsc": "^5.7.3",
+    "avsc": "^5.7.4",
     "benchmarkify": "^3.0.0",
     "bunyan": "^1.8.15",
-    "cbor-x": "^0.9.4",
+    "cbor-x": "^1.2.0",
     "coveralls": "^3.1.1",
     "dd-trace": "^0.36.3",
-    "debug": "^4.3.3",
-    "dotenv": "^10.0.0",
-    "eslint": "^8.6.0",
-    "eslint-config-prettier": "^8.3.0",
+    "debug": "^4.3.4",
+    "dotenv": "^16.0.0",
+    "eslint": "^8.13.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-security": "^1.4.0",
     "etcd3": "^1.1.0",
-    "event-loop-stats": "^1.3.0",
+    "event-loop-stats": "^1.4.1",
     "fakerator": "^0.3.5",
     "gc-stats": "^1.4.0",
-    "ioredis": "^4.28.2",
+    "gen-esm-wrapper": "^1.1.3",
+    "ioredis": "^4.28.5",
     "jaeger-client": "^3.19.0",
-    "jest": "^27.4.7",
-    "jest-cli": "^27.4.7",
-    "jest-diff": "^27.4.6",
-    "joi": "^17.5.0",
+    "jest": "^27.5.1",
+    "jest-cli": "^27.5.1",
+    "jest-diff": "^27.5.1",
+    "joi": "^17.6.0",
     "kafka-node": "^5.0.0",
-    "lockfile-lint": "^4.6.2",
-    "log4js": "^6.3.0",
-    "moleculer-repl": "^0.6.6",
-    "mqtt": "^4.3.4",
-    "msgpack5": "^5.3.2",
-    "nats": "^2.4.0",
-    "node-nats-streaming": "^0.0.51",
+    "lockfile-lint": "^4.7.4",
+    "log4js": "^6.4.5",
+    "moleculer-repl": "^0.7.0",
+    "mqtt": "^4.3.7",
+    "msgpack5": "^6.0.0",
+    "nats": "^2.6.1",
+    "node-nats-streaming": "^0.3.2",
     "nodemon": "^2.0.15",
     "notepack.io": "^2.3.0",
     "npm-check": "^5.9.2",
-    "pino": "^6.13.3",
-    "prettier": "^2.5.1",
+    "pino": "^6.14.0",
+    "prettier": "^2.6.2",
     "protobufjs": "^6.11.2",
     "redlock": "^4.2.0",
     "rhea-promise": "^2.1.0",
-    "supertest": "^6.1.6",
-    "thrift": "^0.12.0",
-    "ts-node": "^10.4.0",
-    "tsd": "^0.13.1",
-    "typescript": "^4.5.4",
+    "supertest": "^6.2.2",
+    "thrift": "^0.16.0",
+    "ts-node": "^10.7.0",
+    "tsd": "^0.20.0",
+    "typescript": "^4.6.3",
     "v8-natives": "^1.2.5",
-    "winston": "^3.3.3",
+    "winston": "^3.7.2",
     "winston-context": "^0.0.7"
   },
   "dependencies": {
