@@ -28,11 +28,11 @@ declare namespace Moleculer {
 	}
 
 	interface LoggerBindings {
-  		nodeID: string;
-  		ns: string;
-  		mod: string;
-  		svc: string;
-  		ver: string | void;
+		nodeID: string;
+		ns: string;
+		mod: string;
+		svc: string;
+		ver: string | void;
 	}
 
 	class LoggerInstance {
@@ -77,7 +77,7 @@ declare namespace Moleculer {
 
 	interface TracerOptions {
 		enabled?: boolean;
-		exporter?: string | TracerExporterOptions | Array<TracerExporterOptions|string> | null;
+		exporter?: string | TracerExporterOptions | Array<TracerExporterOptions | string> | null;
 		sampling?: {
 			rate?: number | null;
 			tracesPerSecond?: number | null;
@@ -205,13 +205,13 @@ declare namespace Moleculer {
 	}
 
 	namespace TracerExporters {
-		class Base extends BaseTraceExporter {}
-		class Console extends BaseTraceExporter {}
-		class Datadog extends BaseTraceExporter {}
-		class Event extends BaseTraceExporter {}
-		class EventLegacy extends BaseTraceExporter {}
-		class Jaeger extends BaseTraceExporter {}
-		class Zipkin extends BaseTraceExporter {}
+		class Base extends BaseTraceExporter { }
+		class Console extends BaseTraceExporter { }
+		class Datadog extends BaseTraceExporter { }
+		class Event extends BaseTraceExporter { }
+		class EventLegacy extends BaseTraceExporter { }
+		class Jaeger extends BaseTraceExporter { }
+		class Zipkin extends BaseTraceExporter { }
 	}
 
 	interface MetricsReporterOptions {
@@ -223,7 +223,7 @@ declare namespace Moleculer {
 		enabled?: boolean;
 		collectProcessMetrics?: boolean;
 		collectInterval?: number;
-		reporter?: string | MetricsReporterOptions | Array<MetricsReporterOptions|string> | null;
+		reporter?: string | MetricsReporterOptions | Array<MetricsReporterOptions | string> | null;
 		defaultBuckets?: Array<number>;
 		defaultQuantiles?: Array<number>;
 		defaultMaxAgeSeconds?: number;
@@ -331,11 +331,11 @@ declare namespace Moleculer {
 	}
 
 	namespace MetricTypes {
-		class Base extends BaseMetric {}
-		class Counter extends CounterMetric {}
-		class Gauge extends GaugeMetric {}
-		class Histogram extends HistogramMetric {}
-		class Info extends InfoMetric {}
+		class Base extends BaseMetric { }
+		class Counter extends CounterMetric { }
+		class Gauge extends GaugeMetric { }
+		class Histogram extends HistogramMetric { }
+		class Info extends InfoMetric { }
 	}
 
 	interface BaseMetricOptions {
@@ -411,13 +411,13 @@ declare namespace Moleculer {
 	}
 
 	namespace MetricReporters {
-		class Base extends MetricBaseReporter {}
-		class Console extends MetricBaseReporter {}
-		class CSV extends MetricBaseReporter {}
-		class Event extends MetricBaseReporter {}
-		class Datadog extends MetricBaseReporter {}
-		class Prometheus extends MetricBaseReporter {}
-		class StatsD extends MetricBaseReporter {}
+		class Base extends MetricBaseReporter { }
+		class Console extends MetricBaseReporter { }
+		class CSV extends MetricBaseReporter { }
+		class Event extends MetricBaseReporter { }
+		class Datadog extends MetricBaseReporter { }
+		class Prometheus extends MetricBaseReporter { }
+		class StatsD extends MetricBaseReporter { }
 	}
 
 	interface BulkheadOptions {
@@ -451,7 +451,7 @@ declare namespace Moleculer {
 		error?: string | ActionHookError | Array<string | ActionHookError>;
 	}
 
-	interface RestSchema{
+	interface RestSchema {
 		path?: string,
 		method?: 'GET' | 'POST' | 'DELETE' | 'PUT' | 'PATCH',
 		fullPath?: string,
@@ -504,7 +504,7 @@ declare namespace Moleculer {
 		metadata: GenericObject;
 
 		ipList: Array<string>;
-		port: number| null;
+		port: number | null;
 		hostname: string | null;
 		udpAddress: string | null;
 
@@ -521,7 +521,7 @@ declare namespace Moleculer {
 		disconnected(): void;
 	}
 
-	class Context<P = unknown, M extends object = {}> {
+	class Context<P = unknown, M extends object = {}, L = GenericObject> {
 		constructor(broker: ServiceBroker, endpoint: Endpoint);
 		id: string;
 		broker: ServiceBroker;
@@ -546,7 +546,7 @@ declare namespace Moleculer {
 		needAck: boolean | null;
 		ackID: string | null;
 
-		locals: GenericObject;
+		locals: L;
 
 		level: number;
 
@@ -600,7 +600,7 @@ declare namespace Moleculer {
 		[name: string]: any;
 	}
 
-	type ServiceEventLegacyHandler = (payload: any,	sender: string,	eventName: string, ctx: Context) => void;
+	type ServiceEventLegacyHandler = (payload: any, sender: string, eventName: string, ctx: Context) => void;
 
 	type ServiceEventHandler = (ctx: Context) => void;
 
@@ -616,17 +616,17 @@ declare namespace Moleculer {
 
 	type ServiceEvents = { [key: string]: ServiceEventHandler | ServiceEventLegacyHandler | ServiceEvent };
 
-	type ServiceMethods = {	[key: string]: (...args: any[]) => any } & ThisType<Service>;
+	type ServiceMethods = { [key: string]: (...args: any[]) => any } & ThisType<Service>;
 
 	type CallMiddlewareHandler = (actionName: string, params: any, opts: CallingOptions) => Promise<any>;
 	type Middleware = {
 		[name: string]:
-			| ((handler: ActionHandler, action: ActionSchema) => any)
-			| ((handler: ActionHandler, event: ServiceEvent) => any)
-			| ((handler: ActionHandler) => any)
-			| ((service: Service) => any)
-			| ((broker: ServiceBroker) => any)
-			| ((handler: CallMiddlewareHandler) => CallMiddlewareHandler)
+		| ((handler: ActionHandler, action: ActionSchema) => any)
+		| ((handler: ActionHandler, event: ServiceEvent) => any)
+		| ((handler: ActionHandler) => any)
+		| ((service: Service) => any)
+		| ((broker: ServiceBroker) => any)
+		| ((handler: CallMiddlewareHandler) => CallMiddlewareHandler)
 	}
 
 	type MiddlewareInit = (broker: ServiceBroker) => Middleware;
@@ -687,7 +687,7 @@ declare namespace Moleculer {
 		[name: string]: any;
 	}
 
-	type ServiceAction = <T = Promise<any>,	P extends GenericObject = GenericObject>(params?: P, opts?: CallingOptions) => T;
+	type ServiceAction = <T = Promise<any>, P extends GenericObject = GenericObject>(params?: P, opts?: CallingOptions) => T;
 
 	interface ServiceActions {
 		[name: string]: ServiceAction;
@@ -695,7 +695,7 @@ declare namespace Moleculer {
 
 	interface WaitForServicesResult {
 		services: string[];
-		statuses: Array<{ name: string; available: boolean}>;
+		statuses: Array<{ name: string; available: boolean }>;
 	}
 
 	class Service<S = ServiceSettingSchema> implements ServiceSchema {
@@ -781,7 +781,7 @@ declare namespace Moleculer {
 		options: DiscovererOptions
 	}
 
-	interface DiscovererOptions extends GenericObject  {
+	interface DiscovererOptions extends GenericObject {
 		heartbeatInterval?: number;
 		heartbeatTimeout?: number;
 		disableHeartbeatChecks?: boolean;
@@ -860,9 +860,6 @@ declare namespace Moleculer {
 		hotReload?: boolean | HotReloadOptions;
 
 		middlewares?: Array<Middleware | string>;
-
-		replCommands?: Array<GenericObject>;
-		replDelimiter?: string;
 
 		metadata?: GenericObject;
 
@@ -943,7 +940,7 @@ declare namespace Moleculer {
 		caller?: string;
 	}
 
-	interface MCallCallingOptions extends CallingOptions{
+	interface MCallCallingOptions extends CallingOptions {
 		settled?: boolean;
 	}
 
@@ -988,14 +985,8 @@ declare namespace Moleculer {
 
 	interface ServiceSearchObj {
 		name: string;
-		version?: string|number;
+		version?: string | number;
 	}
-
-	interface MoleculerRepl extends Vorpal{
-		removeIfExist(command:string): void;
-	}
-
-
 
 	namespace Loggers {
 		class Base {
@@ -1046,8 +1037,6 @@ declare namespace Moleculer {
 
 		start(): Promise<void>;
 		stop(): Promise<void>;
-
-		repl(): MoleculerRepl;
 
 		errorHandler(err: Error, info: GenericObject): void;
 
@@ -1235,9 +1224,9 @@ declare namespace Moleculer {
 			get(key: string): Promise<null | GenericObject>;
 			getWithTTL(key: string): Promise<null | GenericObject>;
 			set(key: string, data: any, ttl?: number): Promise<any>;
-			del(key: string|Array<string>): Promise<any>;
-			clean(match?: string|Array<string>): Promise<any>;
-			getCacheKey(actionName: string, params: object, meta: object, keys: Array<string> | null) : string;
+			del(key: string | Array<string>): Promise<any>;
+			clean(match?: string | Array<string>): Promise<any>;
+			getCacheKey(actionName: string, params: object, meta: object, keys: Array<string> | null): string;
 			defaultKeygen(actionName: string, params: object | null, meta: object | null, keys: Array<string> | null): string;
 		}
 
@@ -1290,32 +1279,32 @@ declare namespace Moleculer {
 		convertSchemaToMoleculer(schema: any): GenericObject;
 	}
 
-	class Validator extends BaseValidator {} // deprecated
+	class Validator extends BaseValidator { } // deprecated
 
 	abstract class BaseStrategy {
-		constructor(registry:ServiceRegistry, broker:ServiceBroker, opts?:object);
+		constructor(registry: ServiceRegistry, broker: ServiceBroker, opts?: object);
 		select(list: any[], ctx?: Context): Endpoint;
 	}
 
 	type ValidatorNames = "Fastest"
 
-	class RoundRobinStrategy extends BaseStrategy {}
-	class RandomStrategy extends BaseStrategy {}
-	class CpuUsageStrategy extends BaseStrategy {}
-	class LatencyStrategy extends BaseStrategy {}
-	class ShardStrategy extends BaseStrategy {}
+	class RoundRobinStrategy extends BaseStrategy { }
+	class RandomStrategy extends BaseStrategy { }
+	class CpuUsageStrategy extends BaseStrategy { }
+	class LatencyStrategy extends BaseStrategy { }
+	class ShardStrategy extends BaseStrategy { }
 
 	namespace Strategies {
-		class Base extends BaseStrategy {}
-		class RoundRobin extends RoundRobinStrategy {}
-		class Random extends RandomStrategy {}
-		class CpuUsage extends CpuUsageStrategy {}
-		class Latency extends LatencyStrategy {}
-		class Shard extends ShardStrategy {}
+		class Base extends BaseStrategy { }
+		class RoundRobin extends RoundRobinStrategy { }
+		class Random extends RandomStrategy { }
+		class CpuUsage extends CpuUsageStrategy { }
+		class Latency extends LatencyStrategy { }
+		class Shard extends ShardStrategy { }
 	}
 
 	abstract class BaseDiscoverer {
-		constructor(opts?:DiscovererOptions);
+		constructor(opts?: DiscovererOptions);
 
 		transit?: Transit;
 		localNode?: BrokerNode;
@@ -1333,23 +1322,23 @@ declare namespace Moleculer {
 		beat(): Promise<void>;
 		checkRemoteNodes(): void;
 		checkOfflineNodes(): void;
-		heartbeatReceived(nodeID:string, payload:GenericObject): void;
-		processRemoteNodeInfo(nodeID:string, payload:GenericObject): BrokerNode;
+		heartbeatReceived(nodeID: string, payload: GenericObject): void;
+		processRemoteNodeInfo(nodeID: string, payload: GenericObject): BrokerNode;
 		sendHeartbeat(): Promise<void>;
 		discoverNode(nodeID: string): Promise<BrokerNode | void>;
 		discoverAllNodes(): Promise<BrokerNode[] | void>;
 		localNodeReady(): Promise<void>;
 		sendLocalNodeInfo(nodeID: string): Promise<void>;
 		localNodeDisconnected(): Promise<void>;
-		remoteNodeDisconnected(nodeID:string, isUnexpected:boolean): void;
+		remoteNodeDisconnected(nodeID: string, isUnexpected: boolean): void;
 
 	}
 
 	namespace Discoverers {
-		class Base extends BaseDiscoverer {}
-		class Local extends BaseDiscoverer {}
-		class Redis extends BaseDiscoverer {}
-		class Etcd3 extends BaseDiscoverer {}
+		class Base extends BaseDiscoverer { }
+		class Local extends BaseDiscoverer { }
+		class Redis extends BaseDiscoverer { }
+		class Etcd3 extends BaseDiscoverer { }
 	}
 
 	interface ValidatorOptions {
@@ -1358,12 +1347,12 @@ declare namespace Moleculer {
 	}
 
 	namespace Validators {
-		class Base extends BaseValidator {}
-		class Fastest extends BaseValidator {}
+		class Base extends BaseValidator { }
+		class Fastest extends BaseValidator { }
 	}
 
 	namespace Transporters {
-		class Base extends Transporter {}
+		class Base extends Transporter { }
 		class Fake extends Base { }
 		class NATS extends Base { }
 		class MQTT extends Base { }
@@ -1496,10 +1485,10 @@ declare namespace Moleculer {
 	}
 
 	interface ActionCatalogListOptions {
-		onlyLocal?:boolean;
-		onlyAvailable?:boolean;
-		skipInternal?:boolean;
-		withEndpoints?:boolean;
+		onlyLocal?: boolean;
+		onlyAvailable?: boolean;
+		skipInternal?: boolean;
+		withEndpoints?: boolean;
 	}
 
 	class ServiceRegistry {
@@ -1536,85 +1525,6 @@ declare namespace Moleculer {
 	const CIRCUIT_CLOSE: string;
 	const CIRCUIT_HALF_OPEN: string;
 	const CIRCUIT_OPEN: string;
-
-	class Vorpal {
-		parse(argv: ReadonlyArray<string>): this;
-		delimiter(value: string): this;
-		show(): this;
-		hide(): this;
-		find(command: string): Vorpal.Command;
-		exec(command: string): Promise<{}>;
-		execSync(command: string): Promise<{}>;
-		log(value: string, ...values: string[]): this;
-		history(id: string): this;
-		localStorage(id: string): object;
-		help(value: (cmd: string) => string): this;
-		pipe(value: (stdout: string) => string): this;
-		use(extension: Vorpal.Extension): this;
-		catch(command: string, description?: string): Vorpal.Catch;
-		command(command: string, description?: string): Vorpal.Command;
-		version(version: string): this;
-		sigint(value: () => void): this;
-		ui: Vorpal.UI;
-		activeCommand: Vorpal.CommandInstance;
-	}
-
-	namespace Vorpal {
-		interface Args {
-			[key: string]: any;
-			options: {
-				[key: string]: any;
-			};
-		}
-
-		interface PromptObject {
-			[key: string]: any;
-		}
-
-		type Action = (args: Args) => Promise<void>;
-		type Cancel = () => void;
-
-		class Command {
-			_name: string;
-			_fn: Action;
-			_cancel: Cancel | undefined;
-			alias(command: string): this;
-			parse(value: (command: string, args: Args) => string): this;
-			option(option: string, description: string, autocomplete?: ReadonlyArray<string>): this;
-			types(types: { string?: ReadonlyArray<string> }): this;
-			hidden(): this;
-			remove(): this;
-			help(value: (args: Args) => void): this;
-			validate(value: (args: Args) => boolean | string): this;
-			autocomplete(values: ReadonlyArray<string> | { data: () => Promise<ReadonlyArray<string>> }): this;
-			action(action: Action): this;
-			cancel(cancel: Cancel): this;
-			allowUnknownOptions(): this;
-		}
-
-		class Catch extends Command { }
-
-		class Extension { }
-
-		class UI {
-			delimiter(text?: string): string;
-			input(text?: string): string;
-			imprint(): void;
-			submit(command: string): string;
-			cancel(): void;
-			redraw: {
-				(text: string, ...texts: string[]): void;
-				clear(): void;
-				done(): void;
-			};
-		}
-
-		class CommandInstance {
-			log(value: string, ...values: string[]): void;
-			prompt(prompt: object | ReadonlyArray<object>): Promise<PromptObject>;
-			delimiter(value: string): void;
-		}
-	}
 
 	namespace Utils {
 		function isFunction(func: unknown): func is Function;
