@@ -6,5 +6,20 @@ module.exports = {
 	singleQuote: false,
 	semi: true,
 	bracketSpacing: true,
-	arrowParens: "avoid"
+	arrowParens: "avoid",
+	overrides: [
+		{
+			files: "*.md",
+			options: {
+				useTabs: false
+			}
+		},
+		{
+			files: "*.json",
+			options: {
+				tabWidth: 2,
+				useTabs: false
+			}
+		}
+	]
 };
