@@ -174,7 +174,7 @@ class Registry {
 
 			this.nodes.localNode.services.push(service);
 
-			this.regenerateLocalRawInfo(this.broker.started);
+			this.regenerateLocalRawInfo(true);
 
 			this.logger.info(`'${svc.name}' service is registered.`);
 
@@ -457,7 +457,7 @@ class Registry {
 			"seq",
 			"metadata"
 		]);
-		if (this.broker.started) rawInfo.services = this.services.getLocalNodeServices();
+		if (this.broker.started || incSeq) rawInfo.services = this.services.getLocalNodeServices();
 		else rawInfo.services = [];
 
 		// Make to be safety
