@@ -164,6 +164,7 @@ class Transit {
 				});
 
 				if (this.__connectResolve) {
+					this.isReady = true;
 					this.__connectResolve();
 					this.__connectResolve = null;
 				}
@@ -243,9 +244,9 @@ class Transit {
 	 */
 	ready() {
 		if (this.connected) {
-			this.isReady = true;
 			this.metrics.set(METRIC.MOLECULER_TRANSIT_READY, 1);
-			return this.discoverer.localNodeReady();
+			// We do nothing here because INFO packets are sent during the starting process.
+			return;
 		}
 	}
 
@@ -426,7 +427,7 @@ class Transit {
 				"."
 		);
 
-		if (!this.broker.started) {
+		if (this.broker.stopping) {
 			this.logger.warn(
 				`Incoming '${payload.event}' event from '${payload.sender}' node is dropped, because broker is stopped.`
 			);
@@ -464,7 +465,7 @@ class Transit {
 		this.logger.debug(`<= Request '${payload.action}' received from '${payload.sender}' node.`);
 
 		try {
-			if (!this.broker.started) {
+			if (this.broker.stopping) {
 				this.logger.warn(
 					`Incoming '${payload.action}' request from '${payload.sender}' node is dropped because broker is stopped.`
 				);
