@@ -54,8 +54,8 @@ describe("Test circuit breaker", () => {
 		return master1
 			.start()
 			.then(() => slave1.start())
-			.delay(100)
-			.then(() => (clock = lolex.install()));
+			.delay(2000)
+			.then(() => (clock = lolex.install({ shouldClearNativeTimers: true })));
 	});
 
 	afterAll(() => {
